업데이트: Certificate 삭제기능 구현 ver1_작성자만 삭제가능

+++ back/src/routers/certificateRouter.js
@@ -81,11 +81,16 @@ certificateRouter.put(
 
 certificateRouter.delete(
   "/certificates/:certificateId",
-  // login_required,
+  login_required,
   async (req, res, next) => {
+    const userId = req.currentUserId;
     const { certificateId } = req.params;
-    await CertificateModel.findOneAndDelete({ certificateId });
-    return res.status(201).json({ msg: "삭제성공" });
+    const foundCertificate = await CertificateModel.findOne({ certificateId });
+    if (userId === foundCertificate.userId) {
+      await CertificateModel.findOneAndDelete({ certificateId });
+      return res.status(201).json({ msg: "성공적으로 삭제 되었습니다." });
+    }
+    res.status(400).json({ message: "작성자만 삭제할 수 있습니다." });
   }
 );
